Fix CodeNarc violations in matrix-ggplot

+++ matrix-ggplot/src/test/groovy/gg/scale/ScaleSizeIdentityTest.groovy
@@ -19,7 +19,7 @@ class ScaleSizeIdentityTest {
     def scale = new ScaleSizeIdentity()
     assertEquals(5 as BigDecimal, scale.transform(5))
     assertEquals(10 as BigDecimal, scale.transform(10))
-    assertEquals(2.5 as BigDecimal, scale.transform(2.5))
+    assertEquals(2.5, scale.transform(2.5))
   }
 
   @Test
@@ -27,7 +27,7 @@ class ScaleSizeIdentityTest {
     def scale = new ScaleSizeIdentity()
     def result = scale.transform('7.5')
     assertNotNull(result)
-    assertEquals(7.5 as BigDecimal, result)
+    assertEquals(7.5, result)
   }
 
   @Test
@@ -39,48 +39,48 @@ class ScaleSizeIdentityTest {
   @Test
   void testBigDecimalTransform() {
     def scale = new ScaleSizeIdentity()
-    assertEquals(12.75 as BigDecimal, scale.transform(12.75G))
+    assertEquals(12.75, scale.transform(12.75G))
   }
 
   @Test
   void testNegativeValue() {
     def scale = new ScaleSizeIdentity()
-    assertEquals(0.1 as BigDecimal, scale.transform(-5))
-    assertEquals(0.1 as BigDecimal, scale.transform(-0.5))
+    assertEquals(0.1, scale.transform(-5))
+    assertEquals(0.1, scale.transform(-0.5))
   }
 
   @Test
   void testZeroValue() {
     def scale = new ScaleSizeIdentity()
-    assertEquals(0.1 as BigDecimal, scale.transform(0))
-    assertEquals(0.1 as BigDecimal, scale.transform(0.0))
+    assertEquals(0.1, scale.transform(0))
+    assertEquals(0.1, scale.transform(0.0))
   }
 
   @Test
   void testVerySmallPositiveValue() {
     def scale = new ScaleSizeIdentity()
-    assertEquals(0.1 as BigDecimal, scale.transform(0.05))
-    assertEquals(0.1 as BigDecimal, scale.transform(0.001))
+    assertEquals(0.1, scale.transform(0.05))
+    assertEquals(0.1, scale.transform(0.001))
   }
 
   @Test
   void testNonNumericString() {
     def scale = new ScaleSizeIdentity()
-    assertEquals(3.0 as BigDecimal, scale.transform('abc'))
-    assertEquals(3.0 as BigDecimal, scale.transform('not a number'))
+    assertEquals(3.0, scale.transform('abc'))
+    assertEquals(3.0, scale.transform('not a number'))
   }
 
   @Test
   void testEmptyString() {
     def scale = new ScaleSizeIdentity()
-    assertEquals(3.0 as BigDecimal, scale.transform(''))
+    assertEquals(3.0, scale.transform(''))
   }
 
   @Test
   void testBooleanValue() {
     def scale = new ScaleSizeIdentity()
-    assertEquals(3.0 as BigDecimal, scale.transform(true))
-    assertEquals(3.0 as BigDecimal, scale.transform(false))
+    assertEquals(3.0, scale.transform(true))
+    assertEquals(3.0, scale.transform(false))
   }
 
   @Test
